NOTICE about conflict resolution in MEGA Sync
The way MEGA Sync resolves conflicts with the last modified date is _very_ different from traditional schemes and this unexpected behavior can cause many problems for users. Adding a notice to the main readme warning users about this behaviour is the least we can do to prevent loss of data and other issues.